% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\alias{describeMetadata}
\alias{describeResults}
\alias{showMap}
\title{Describe characteristics of Airbnb listings.}
\usage{
describeMetadata(content)

describeResults(content)

showMap(content)
}
\arguments{
\item{content}{The output of \code{\link{searchLocation}}.}
}
\description{
Given output from \code{\link{searchLocation}}, produces
summary charts based on metadata and results.
}
\examples{
describeMetadata(content)

describeResults(content)

showMap(content)

}
