% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailsAdded.R
\docType{data}
\name{detailsAdded}
\alias{detailsAdded}
\title{Listing data from \code{searchLocation} with more detailed 
information merged in using \code{addDetails}.}
\description{
Description.
}
\keyword{data}
