% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictPrice.R
\name{predictPrice}
\alias{predictPrice}
\title{Predict a price for an Airbnb listing using the characteristics of nearby listings.}
\usage{
predictPrice(listingID, maxSample = 500, nfold = 10, nrounds = 1000,
  early_stopping_rounds = 50, max_depth = 4, eta = 0.05,
  listing.detail = NULL, trainData = NULL)
}
\arguments{
\item{listingID}{the ID that Airbnb has assigned to the listing. This can be taken from the end of a listing's
URL. This should be passed as a string (a character vector of length one).}

\item{maxSample}{The maximum number of nearby listings on which to base the price prediction. There is a
tradeoff-a higher value will increase prediction precision, but it will take longer to run.}

\item{nfold}{For advanced users. The number of cross validation folds used by \code{\link{xgboost}}.}

\item{nrounds}{For advanced users. The maximum number of rounds that \code{\link{xgboost}} will run
when finding the optimal number of rounds through cross validation.}

\item{early_stopping_rounds}{For advanced users. This parameter will cause xgboost to stop
training new models after the specified number of rounds if cross validation error does not improve.}

\item{max_depth}{For advanced users. The maximum depth of a tree fitted by \code{\link{xgboost}}.}

\item{eta}{For advanced users. The shrinkage parameter used to control the learning rate for
\code{\link{xgboost}}.}

\item{listing.detail}{a dataset for the particular listing if you've already pulled it using
\code{\link{listingDetails}}. Note that this should be a dataset of only one listing (as in 
\code{\link{listing.detail}}).}

\item{trainData}{a dataset consisting of listings with details if you've already pulled it
using \code{\link{searchLocation}} and \code{\link{addDetails}}.}
}
\description{
Given a listing ID, \code{predictPrice} uses the \code{\link{xgboost}} package to predict a price
for that listing based on its characteristics and data from nearby listings. A listing ID can be found
on the end of the URL for the listing on Airbnb's site. For example, a listing URL for Airbnb
looks like \code{https://www.airbnb.com/rooms/17634206?s=QAXAT6DD}. In this example, the listing
ID is "17634206." Passing this value as a character string will yield a price prediction based
on its characteristics, implying over- or undervaluation.
}
\examples{
data("trainData", package="Rbnb")
data("listing.detail", package="Rbnb")
price.prediction <- Rbnb::predictPrice("12170773",
       listing.detail = listing.detail, trainData = trainData[1:100])

}
