% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchLocation.R
\name{searchLocation}
\alias{searchLocation}
\title{Search for Airbnb listings by location.}
\usage{
searchLocation(location, verbose = TRUE, metadata.only = FALSE,
  client.id = "d306zoyjsyarp7ifhu67rjxn52tv0t20")
}
\arguments{
\item{location}{a string representing the desired search region.}

\item{verbose}{a boolean indicating whether or not to print status updates.}

\item{metadata.only}{a boolean indicating whether or not to return just
metadata (no listings).}

\item{client.id}{a string represting your own Airbnb API key.}
}
\value{
named list containing the serach output.
}
\description{
Given an input value, lets Airbnb parse it, and return listings
in the same region.
}
\examples{
searchLocation("10019")

}
