% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainData.R
\docType{data}
\name{trainData}
\alias{trainData}
\title{trainData}
\description{
Listing data from \code{\link{searchLocation}} with details merged in with
\code{\link{addDetails}} for use in predicting a price for Listing \code{12170773}.
}
\keyword{data}
