% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDetails.R
\name{addDetails}
\alias{addDetails}
\alias{listingDetails}
\alias{mergeDetails}
\alias{getListingDetail}
\title{Pull in detailed listing data to merge with location search data.}
\usage{
listingDetails(listingIDs, client_id = "d306zoyjsyarp7ifhu67rjxn52tv0t20")

mergeDetails(searchResults, details)

getListingDetail(listingID, client_id = "d306zoyjsyarp7ifhu67rjxn52tv0t20")
}
\arguments{
\item{searchData}{a dataset of listings outputted from \code{searchLocation}.}
}
\description{
After using the \code{searchLocation}, you can add greater detail on each listing
using \code{addDetails}. It takes as an input the dataset of listings outputted from
\code{searchLocation} and merges in details based on the listing ID. Note that this process
is very time consuming, and so it may be unwise to pass too large of a listing dataset.

The following is a function that will take in a character vector of
listing IDs and return a dataset. This function does not necessarily need to be used
with \code{locationSearch}.

The following is a function that will take in a character vector of
listing IDs and return a dataset. This function does not necessarily need to be used
with \code{locationSearch}.

The following is a function that will take in a character vector of
listing IDs and return a dataset. This function does not necessarily need to be used
with \code{locationSearch}.
}
\examples{
#content  <- searchLocation("10019")
#searchData <- content$results$data
#listingsWithDetails <- addDetails(searchData)

}
