% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\alias{describeMetadata}
\alias{describeResults}
\alias{showMap}
\title{Describe characteristics of Airbnb listings.}
\usage{
describeMetadata(content, facet = "bedrooms")

describeResults(content)

showMap(content)
}
\arguments{
\item{content}{The output of \code{\link{searchLocation}}.}

\item{facet}{Which facet you'd like to plot. One of c("bedrooms", "beds",
"bathrooms", "room.type", "hosting.amenities").}
}
\description{
Given output from \code{\link{searchLocation}}, produces
summary charts based on metadata and results.
}
\examples{
data("zip10019", package = "Rbnb")
describeMetadata(zip10019)

describeResults(zip10019)

showMap(zip10019)

}
