% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainData.R
\docType{data}
\name{trainData}
\alias{trainData}
\title{Listing data from \code{searchLocation} with details merged in with
\code{addDetails} for use in predicting a price for Listing \code{12170773}.}
\description{
Description.
}
\keyword{data}
